requirements.txt
gradio==5.49.1
langchain==0.3.12
python-dotenv==1.0.1
langchain-core==0.3.28
langchain-openai==0.2.12
langchain-community==0.3.12
langchain-text-splitters==0.3.4
langchain-groq==0.2.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
beautifulsoup4==4.12.3
requests==2.32.3
pypdf==5.1.0
lxml==6.0.2